refactor(DEV-10395): `<PayableStatusChip/>` to separate directory
