+++ packages/sdk-react/src/components/payables/PayableStatusChip/index.ts
@@ -1,0 +1,4 @@
+export {
+  PayableStatusChip,
+  type PayableStatusChipProps,
+} from './PayableStatusChip';
