+++ packages/sdk-react/src/components/payables/PayablesTable/PayablesTable.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 
 import { ScopedCssBaselineContainerClassName } from '@/components/ContainerCssBaseline';
 import { CounterpartCell } from '@/components/payables/PayablesTable/CounterpartCell/CounterpartCell';
-import { PayableStatusChip } from '@/components/payables/PayablesTable/PayableStatusChip';
+import { PayableStatusChip } from '@/components/payables/PayableStatusChip';
 import { PAGE_LIMIT } from '@/constants';
 import { MoniteScopedProviders } from '@/core/context/MoniteScopedProviders';
 import { useCurrencies } from '@/core/hooks/useCurrencies';
